+++ src/main/java/doubleCache/Subject.java
@@ -9,9 +9,9 @@ import lombok.ToString;
 @Getter
 @ToString
 public class Subject {
-    private static long counter = 0L;
+    private static int counter = 0;
 
-    private final long id;
+    private final int id;
     private final String name;
 
     public Subject(String name) {
